reorganize point panels folder
