+++ frontend/farm_designer/plants/add_plant.tsx
@@ -5,7 +5,7 @@ import { svgToUrl } from "../../open_farm/icons";
 import { CropLiveSearchResult, OpenfarmSearch } from "../interfaces";
 import { setDragIcon } from "../map/actions";
 import { getCropHeaderProps, searchForCurrentCrop } from "./crop_info";
-import { DesignerPanel, DesignerPanelHeader } from "./designer_panel";
+import { DesignerPanel, DesignerPanelHeader } from "../designer_panel";
 import { OFSearch } from "../util";
 import { t } from "../../i18next_wrapper";
 import { Panel } from "../panel_header";
