--- frontend/farm_designer/plants/point_inventory.tsx
@@ -1,78 +1,0 @@
-import * as React from "react";
-import { connect } from "react-redux";
-import { PointInventoryItem } from "./point_inventory_item";
-import { Everything } from "../../interfaces";
-import { DesignerNavTabs, Panel } from "../panel_header";
-import {
-  EmptyStateWrapper, EmptyStateGraphic
-} from "../../ui/empty_state_wrapper";
-import { Content } from "../../constants";
-import {
-  DesignerPanel, DesignerPanelContent, DesignerPanelTop
-} from "./designer_panel";
-import { selectAllGenericPointers } from "../../resources/selectors";
-import { TaggedGenericPointer } from "farmbot";
-import { t } from "../../i18next_wrapper";
-import { isAWeed } from "./weeds_inventory";
-
-export interface PointsProps {
-  points: TaggedGenericPointer[];
-  dispatch: Function;
-  hoveredPoint: string | undefined;
-}
-
-interface PointsState {
-  searchTerm: string;
-}
-
-export function mapStateToProps(props: Everything): PointsProps {
-  const { hoveredPoint } = props.resources.consumers.farm_designer;
-  return {
-    points: selectAllGenericPointers(props.resources.index)
-      .filter(x => !x.body.discarded_at)
-      .filter(x => !isAWeed(x.body.name, x.body.meta.type)),
-    dispatch: props.dispatch,
-    hoveredPoint,
-  };
-}
-
-export class RawPoints extends React.Component<PointsProps, PointsState> {
-  state: PointsState = { searchTerm: "" };
-
-  update = ({ currentTarget }: React.SyntheticEvent<HTMLInputElement>) => {
-    this.setState({ searchTerm: currentTarget.value });
-  }
-
-  render() {
-    return <DesignerPanel panelName={"point-inventory"} panel={Panel.Points}>
-      <DesignerNavTabs />
-      <DesignerPanelTop
-        panel={Panel.Points}
-        linkTo={"/app/designer/points/add"}
-        title={t("Add point")}>
-        <input type="text" onChange={this.update}
-          placeholder={t("Search your points...")} />
-      </DesignerPanelTop>
-      <DesignerPanelContent panelName={"points"}>
-        <EmptyStateWrapper
-          notEmpty={this.props.points.length > 0}
-          graphic={EmptyStateGraphic.points}
-          title={t("No points yet.")}
-          text={Content.NO_POINTS}
-          colorScheme={"points"}>
-          {this.props.points
-            .filter(p => p.body.name.toLowerCase()
-              .includes(this.state.searchTerm.toLowerCase()))
-            .map(p => <PointInventoryItem
-              key={p.uuid}
-              navName={"points"}
-              tpp={p}
-              hovered={this.props.hoveredPoint === p.uuid}
-              dispatch={this.props.dispatch} />)}
-        </EmptyStateWrapper>
-      </DesignerPanelContent>
-    </DesignerPanel>;
-  }
-}
-
-export const Points = connect(mapStateToProps)(RawPoints);
