+++ frontend/farm_designer/points/__tests__/point_edit_actions_test.tsx
