+++ frontend/farm_designer/points/__tests__/point_inventory_item_test.tsx
