+++ frontend/farm_designer/points/__tests__/point_inventory_test.tsx
