+++ frontend/farm_designer/points/__tests__/weeds_add_test.tsx
