+++ frontend/farm_designer/points/__tests__/weeds_inventory_test.tsx
