+++ frontend/farm_designer/points/point_inventory_item.tsx
