+++ frontend/farm_designer/tools/add_tool.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { connect } from "react-redux";
 import {
   DesignerPanel, DesignerPanelContent, DesignerPanelHeader
-} from "../plants/designer_panel";
+} from "../designer_panel";
 import { Everything } from "../../interfaces";
 import { t } from "../../i18next_wrapper";
 import { SaveBtn } from "../../ui";
