+++ frontend/farm_designer/zones/zones_inventory.tsx
@@ -8,7 +8,7 @@ import {
 import { Content } from "../../constants";
 import {
   DesignerPanel, DesignerPanelContent, DesignerPanelTop
-} from "../plants/designer_panel";
+} from "../designer_panel";
 import { t } from "../../i18next_wrapper";
 
 export interface ZonesProps {
